Updated package.json scripts to work in linux environments. Moved the test files according to /main hierachy.
