+++ .travis.yml
@@ -5,5 +5,5 @@ node_js:
 before_script:
   - npm install
 script:
-  - npm run test
-  - npm run-script lint
+  - npm test
+  - npm run lint
